cos/queue_ops.rs: colocate the 2 #[ignore]'d micro-benches + tear down empty tx/mod.rs::tests (#984 P3 Phase 7 — final)
Move the last 2 tests from tx/mod.rs::tests:
- cos_exact_drain_throughput_micro_bench
- bench_pop_commit_settle_publish

Both are #[test] #[ignore] micro-benches exercising the cos exact
drain hot path. They belong with cos/queue_ops.rs production code.

Adds: settle_exact_local_scratch_submission_flow_fair,
COS_FLOW_FAIR_BUCKETS imports.

cos_exact_drain_throughput_micro_bench's 'use super::super::worker::'
path was rewritten to crate::afxdp::worker (explicit path required
at the new nesting depth).

**Final cleanup of the inline mod tests block in tx/mod.rs**:
With this PR the inline #[cfg(test)] mod tests { ... } block is
empty. Codex round-1 flagged it as unused-import debris; remove the
block entirely. Also drops 5 stale cfg-test imports
(ExactCoSScratchBuild, drain_exact_local_fifo_items_to_scratch,
drain_exact_local_items_to_scratch_flow_fair,
settle_exact_local_fifo_submission, settle_exact_prepared_fifo_submission)
that were the bench-only consumers.

Closes #984 P3. tx/mod.rs went from 10,899 lines pre-refactor
(P3 Phase 0) → 167 lines now (98.5% reduction). All 184 tests
moved across 11 sibling files (rings, transmit, tx_completion,
cross_binding, cos_classify, ecn, queue_service, admission,
flow_hash, builders, token_bucket, queue_ops).

cargo test --bins 865/0/2.
